requirements.txt
azure-ai-projects<=1.0.0b7
python-dotenv<=1.1.0
azure-identity<=1.21.0
streamlit<=1.44.0